{stream,transport}: fix connection overwriting when a client uses the same port to connect.
The server's transport maintains an `addrToConn` map based only on the client's IP-Port. When a client connects to different ports on the server using the same local IP-port, it causes the connections in the server transport's `addrToConn` map to be overwritten, leading to a "Can't find conn by addr" error in the streaming service.

So now, the combination of the client's IP-Port and the server's IP-Port is used as the `addrToConn` map key.